requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.117
botocore==1.29.117
bravado==11.0.3
bravado-core==5.17.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
h5py==3.8.0
idna==3.4
imageio==2.27.0
isoduration==20.11.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
kiwisolver==1.4.4
lightning-utilities==0.8.0
lit==16.0.1
MarkupSafe==2.1.2
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
neptune-client==1.1.1
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
psutil==5.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==1.9.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
swagger-spec-validator==3.0.3
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
webcolors==1.13
websocket-client==1.5.1
yarl==1.8.2
